MVP Completed
There's a strange error though. It loads the image, it ports the user over to the detail view, but it can't set the text of my (new) text fields to the user text. If you get a chance, take a look at it and tell me what I am doing wrong. It says that it finds "nil" when trying to unwrap an optional, but the only optional is the textfield itself and it's there. I have the outlets hooked up correctly. Gonna try redoing the DetailVC and see what happens. 

@jonbash